requirements.txt
numpy
easyocr
Pillow
keyboard
opencv-python
pyperclip
pyautogui